+++ MoviesApp/App/components/AddMovie.vue
@@ -55,7 +55,6 @@
 				this.showError = false;
 				if (this.movie.title.length > 0 && this.movie.year > 0 && this.movie.director.length > 0 && this.movie.description.length > 0) {
 					this.$emit('save-movie-add');
-					this.closeAddModal();
 				} else {
 					this.showError = true;
 				}
