+++ MoviesApp/App/components/DisplayMovie.vue
